% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_palettes.R
\docType{data}
\name{oz_palettes}
\alias{oz_palettes}
\title{An Australian palette generator}
\format{An object of class \code{list} of length 6.}
\usage{
oz_palettes
}
\arguments{
\item{n}{Number of colors desired.
If omitted, uses all colours.}

\item{name}{Name of desired palette. Choices are:
\code{grandis}, \code{flavolineata},  \code{whitei},
\code{picta}, \code{virgo},  \code{ngadju}}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
An Australian palette generator
}
\keyword{colours}
