% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRoz.R
\name{colRoz_pal}
\alias{colRoz_pal}
\title{ColRoz palette generator}
\usage{
colRoz_pal(name, n, type = c("discrete", "continuous", "paired"))
}
\arguments{
\item{n:}{Number of colors desired If omitted, uses all colours.}

\item{name:}{Name of color palette desired.}

\item{type:}{Either "discrete", "continuous", or "paired".}
}
\description{
This function allows you to generate color palettes based on Australian scenes
}
\examples{
colRoz_pal(name = "grandis")
For continuous gradients, you can also use colour_fill_gradientn() in ggplot

}
\keyword{color}
\keyword{palettes}
