% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_palettes.R
\docType{data}
\name{oz_palettes}
\alias{oz_palettes}
\title{An Australian palette generator}
\format{An object of class \code{list} of length 21.}
\usage{
oz_palettes
}
\arguments{
\item{n}{Number of colors desired.
If omitted, uses all colours.}

\item{name}{Name of desired palette. Choices are:
\code{BottleRocket1}, \code{BottleRocket2},  \code{Rushmore1},
\code{Royal1}, \code{Royal2},  \code{Zissou1}, \code{Darjeeling1},
\code{Darjeeling2},  \code{Chevalier1} , \code{FantasticFox1} ,
\code{Moonrise1}, \code{Moonrise2}, \code{Moonrise3}, \code{Cavalcanti1},
\code{GrandBudapest1}, \code{GrandBudapest2}, \code{IsleofDogs1}, \code{IsleofDogs2}}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
An Australian palette generator
}
\keyword{colours}
