% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_palettes.R
\docType{data}
\name{oz_palettes}
\alias{oz_palettes}
\title{Palettes inspired by Australia}
\format{An object of class \code{list} of length 5.}
\usage{
oz_palettes
}
\value{
A vector of colours.
}
\description{
Palettes stored in nested lists according to theme.
Choices of palettes are:
  \code{warramaba}: palettes for matchstick grasshoppers (Warramaba)
  \code{lizards}: palettes for Australian lizards,
  \code{landscape}: palettes for Australian landscapes,
  \code{birds}: palettes for Australian birds,
  \code{frogs}: palettes for Australian frogs
}
\examples{
# Call themed sets of palettes
oz_palettes[["lizards"]] # or
oz_palettes$lizards

# Call individual palettes using square brackets
oz_palettes[["lizards"]][["c.decresii"]] # or
oz_palettes$lizards$c.decresii

}
\keyword{colours}
\keyword{palette}
