% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_palette.R
\name{print_palette}
\alias{print_palette}
\title{Function to print color palettes}
\usage{
print_palette(x, ...)
}
\arguments{
\item{n:}{Number of colors desired If omitted, uses all colours.}

\item{name:}{Name of color palette desired.}
}
\description{
Function to print color palettes
}
\examples{
print_palette(colRoz_pal("grandis"))

}
